--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,3 +1,4 @@
+use ckb_vm::decoder::Decoder;
 use ckb_vm::{instructions::instruction_length, Bytes, CoreMachine, Error, Memory, Register, SupportMachine};
 use std::borrow::Cow;
 use std::cell::RefCell;
@@ -102,10 +103,16 @@ pub struct PProfLogger {
     tree_node: Rc<RefCell<PProfRecordTreeNode>>,
     ra_dict: HashMap<u64, Rc<RefCell<PProfRecordTreeNode>>>,
     output_filename: String,
+    cache: HashMap<u64, String>,
+    decoder: Decoder,
 }
 
 impl PProfLogger {
-    pub fn new(filename: String, output_filename: &str) -> Result<Self, Box<dyn std::error::Error>> {
+    pub fn new<R: Register>(
+        filename: String,
+        output_filename: &str,
+        isa: u8,
+    ) -> Result<Self, Box<dyn std::error::Error>> {
         let file = std::fs::File::open(filename)?;
         let mmap = unsafe { memmap::Mmap::map(&file)? };
         let object = object::File::parse(&*mmap)?;
@@ -117,7 +124,25 @@ impl PProfLogger {
             tree_node: tree_root,
             ra_dict: HashMap::new(),
             output_filename: String::from(output_filename),
+            cache: Default::default(),
+            decoder: ckb_vm::decoder::build_decoder::<R>(isa),
         })
+    }
+
+    pub fn gen_tag_string(&mut self, addr: u64) -> String {
+        let value = self.cache.get(&addr);
+        if value.is_some() {
+            value.unwrap().clone()
+        } else {
+            let loc = self.atsl_context.find_location(addr).unwrap();
+            let loc_string = sprint_loc_file(&loc);
+            let frame_iter = self.atsl_context.find_frames(addr).unwrap();
+            let fun_string = sprint_fun(frame_iter);
+            let tag_string = format!("{}:{}", loc_string, fun_string);
+            // update cache
+            self.cache.insert(addr, tag_string.clone());
+            tag_string
+        }
     }
 }
 
@@ -126,19 +151,13 @@ impl<'a, R: Register, M: Memory<REG = R>, Inner: ckb_vm::machine::SupportMachine
 {
     fn on_step(&mut self, machine: &mut ckb_vm::machine::DefaultMachine<'a, Inner>) {
         let pc = machine.pc().to_u64();
-        let mut decoder = ckb_vm::decoder::build_decoder::<R>(machine.isa());
-        let inst = decoder.decode(machine.memory_mut(), pc).unwrap();
-        let inst_length = instruction_length(inst) as u64;
+        let inst = self.decoder.decode(machine.memory_mut(), pc).unwrap();
         let opcode = ckb_vm::instructions::extract_opcode(inst);
         let cycles = machine.instruction_cycle_func().as_ref().map(|f| f(inst)).unwrap_or(0);
         self.tree_node.borrow_mut().cycles += cycles;
 
         let once = |s: &mut Self, addr: u64, link: u64| {
-            let loc = s.atsl_context.find_location(addr).unwrap();
-            let loc_string = sprint_loc_file(&loc);
-            let frame_iter = s.atsl_context.find_frames(addr).unwrap();
-            let fun_string = sprint_fun(frame_iter);
-            let tag_string = format!("{}:{}", loc_string, fun_string);
+            let tag_string = s.gen_tag_string(addr);
             let chd = Rc::new(RefCell::new(PProfRecordTreeNode {
                 name: tag_string,
                 parent: Some(s.tree_node.clone()),
@@ -151,6 +170,7 @@ impl<'a, R: Register, M: Memory<REG = R>, Inner: ckb_vm::machine::SupportMachine
         };
 
         if opcode == ckb_vm::instructions::insts::OP_JAL {
+            let inst_length = instruction_length(inst) as u64;
             let inst = ckb_vm::instructions::Utype(inst);
             // The standard software calling convention uses x1 as the return address register and x5 as an alternate
             // link register.
@@ -161,32 +181,38 @@ impl<'a, R: Register, M: Memory<REG = R>, Inner: ckb_vm::machine::SupportMachine
             }
         };
         if opcode == ckb_vm::instructions::insts::OP_JALR {
+            let inst_length = instruction_length(inst) as u64;
             let inst = ckb_vm::instructions::Itype(inst);
             let base = machine.registers()[inst.rs1()].to_u64();
             let addr = base.wrapping_add(inst.immediate_s() as u64) & 0xfffffffffffffffe;
             let link = pc + inst_length;
-            if self.ra_dict.contains_key(&addr) {
-                self.tree_node = self.ra_dict.get(&addr).unwrap().clone();
+            let value = self.ra_dict.get(&addr);
+            if value.is_some() {
+                self.tree_node = value.unwrap().clone();
             } else {
                 once(self, addr, link);
             }
         };
         if opcode == ckb_vm::instructions::insts::OP_FAR_JUMP_ABS {
+            let inst_length = instruction_length(inst) as u64;
             let inst = ckb_vm::instructions::Utype(inst);
             let addr = (inst.immediate_s() as u64) & 0xfffffffffffffffe;
             let link = pc + inst_length;
-            if self.ra_dict.contains_key(&addr) {
-                self.tree_node = self.ra_dict.get(&addr).unwrap().clone();
+            let value = self.ra_dict.get(&addr);
+            if value.is_some() {
+                self.tree_node = value.unwrap().clone();
             } else {
                 once(self, addr, link);
             }
         }
         if opcode == ckb_vm::instructions::insts::OP_FAR_JUMP_REL {
+            let inst_length = instruction_length(inst) as u64;
             let inst = ckb_vm::instructions::Utype(inst);
             let addr = pc.wrapping_add(inst.immediate_s() as u64) & 0xfffffffffffffffe;
             let link = pc + inst_length;
-            if self.ra_dict.contains_key(&addr) {
-                self.tree_node = self.ra_dict.get(&addr).unwrap().clone();
+            let value = self.ra_dict.get(&addr);
+            if value.is_some() {
+                self.tree_node = value.unwrap().clone();
             } else {
                 once(self, addr, link);
             }
@@ -216,11 +242,12 @@ pub fn quick_start(fl_bin: &str, fl_arg: Vec<&str>, output_filename: &str) -> Re
     let default_machine_builder = ckb_vm::DefaultMachineBuilder::new(default_core_machine)
         .instruction_cycle_func(Box::new(cost_model::instruction_cycles));
     let default_machine = default_machine_builder.build();
-    let pprof_logger = PProfLogger::new(String::from(fl_bin), output_filename).map_err(|e| {
-        let mut stderr = std::io::stderr();
-        writeln!(&mut stderr, "error while loading file {}: {:?}", fl_bin, e).expect("write stderr");
-        Error::Unexpected
-    })?;
+    let pprof_logger =
+        PProfLogger::new::<u64>(String::from(fl_bin), output_filename, default_machine.isa()).map_err(|e| {
+            let mut stderr = std::io::stderr();
+            writeln!(&mut stderr, "error while loading file {}: {:?}", fl_bin, e).expect("write stderr");
+            Error::Unexpected
+        })?;
     let pprof_func_provider = Box::new(pprof_logger);
     let mut machine = machine::PProfMachine::new(default_machine, pprof_func_provider);
     let mut args = vec![fl_bin.to_string().into()];