It can reduce cycles of fib (20) from 5.37s to 3.17s.

```
./target/release/ckb-vm-pprof --bin res/fib > /dev/null  
5.37s user 2.72s system 92% cpu 8.718 total

./target/release/ckb-vm-pprof --bin res/fib > /dev/null  
3.17s user 2.77s system 90% cpu 6.575 total
```
Although it is an extreme case.